Remove old documentapi protocols (Java).
